requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
blessed==1.19.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
codefind==0.1.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
defusedxml==0.7.1
Django==4.0.5
django-environ==0.9.0
django-filter==22.1
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
djoser==2.0.5
drf-yasg==1.20.0
environ==1.0
Faker==13.15.1
idna==3.3
importlib-metadata==4.12.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jurigged==0.5.2
Markdown==3.3.7
MarkupSafe==2.1.1
mysql==0.0.3
mysqlclient==2.1.1
oauthlib==3.2.0
ovld==0.3.2
packaging==21.3
Pillow==9.1.1
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.1
PyYAML==6.0
render==1.0.0
requests==2.28.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.9
utils==1.0.1
watchdog==2.1.9
wcwidth==0.2.5
zipp==3.8.0
django-redis==5.2.0
dotmap==1.3.30
rollbar==0.16.3
python-http-client==3.3.7
